Issue #3060511: (doc) Enhance doc comments on features_drush_help().
Return type 'void' is technically equivalent with NULL, but conceptually they are different.
With "@return *|null" in the doc, an IDE inspection tool might require an explicit "return NULL;" at the end of the function.
With "@return *|void" in the doc, an IDE inspection tool might accept a missing return statement, which means implicitly returning NULL.

As a follow-up, we should change this to "@return *|null", and add a "return NULL;" at the end of the function.