COND: Make sure that $restore_states is not undefined, in _features_restore().
The $restore_states variable is defined if $op is one of 'revert' or 'rebuild'.
It is undefined if $op is 'disable' or 'enable'.

The block where we are adding the condition would not do anything if $restore_states is empty or undefined.
This means adding this condition is neutral to behavior, but may prevent notices and warnings.

In any case it prevents inspection warnings in the IDE.